Renamed symbtrdataextractor/symbtrreader.py -> symbtrdataextractor/SymbTrReader.py
