+++ symbtrdataextractor/SymbTrReader.py
